--- a/test/golang/cases_and_options/relative_path/a.thrift
+++ b/test/golang/cases_and_options/relative_path/a.thrift
@@ -1,0 +1,7 @@
+include "b.thrift"
+
+namespace * a
+
+struct A {
+    b.B ab
+}