… parser/Makefile
… parser/parser_test.go
… parser/thrift.peg.go
… parser/token_test.go
… parser/util.go
… test/golang/cases_and_options/relative_path/b.thrift
… test/golang/cases_and_options/relative_path/included/a.thrift
… test/golang/cases_and_options/relative_path/included/b.thrift