Replace the original tool-generated parser by a handcrafted recursive descent parser to improve performance and maintainability.

```
name               old time/op    new time/op    delta
Parse-8              18.7ms ± 2%     3.6ms ± 5%  -80.60%  (p=0.008 n=5+5)
ParseParallel-8      5.03ms ±13%    1.18ms ± 5%  -76.58%  (p=0.008 n=5+5)
Generate-8            336ms ± 3%     334ms ± 2%     ~     (p=0.841 n=5+5)
All-8                 338ms ± 1%     333ms ± 1%     ~     (p=0.151 n=5+5)
SemanticCheck-8       443µs ± 3%     426µs ± 3%   -3.80%  (p=0.032 n=5+5)
SemanticResolve-8    19.3ms ± 4%     4.1ms ± 1%  -78.69%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
Parse-8              10.7MB ± 0%     0.8MB ± 0%  -92.72%  (p=0.008 n=5+5)
ParseParallel-8      10.7MB ± 0%     0.8MB ± 0%  -92.72%  (p=0.008 n=5+5)
Generate-8           72.8MB ± 0%    72.8MB ± 0%     ~     (p=0.151 n=5+5)
All-8                83.6MB ± 0%    73.6MB ± 0%  -11.93%  (p=0.008 n=5+5)
SemanticCheck-8       169kB ± 0%     169kB ± 0%     ~     (p=0.690 n=5+5)
SemanticResolve-8    10.8MB ± 0%     0.8MB ± 0%  -92.26%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
Parse-8                171k ± 0%       36k ± 0%     ~     (p=0.079 n=4+5)
ParseParallel-8        171k ± 0%       36k ± 0%  -79.16%  (p=0.008 n=5+5)
Generate-8            1.23M ± 0%     1.23M ± 0%   +0.00%  (p=0.040 n=5+5)
All-8                 1.40M ± 0%     1.27M ± 0%   -9.68%  (p=0.008 n=5+5)
SemanticCheck-8         174 ± 0%       174 ± 0%     ~     (all equal)
SemanticResolve-8      173k ± 0%       37k ± 0%     ~     (p=0.079 n=4+5)
```